move-action
This incomplete patch tries to move code around to hide the Set case
of Action. While it's not finished, it at least wraps and moves some
usages thereof. Some of the thought that needs to go into this is e.g.
how to make it textually less invasive and alternative strategies.
