+++ bench/tx-generator/src/Cardano/Benchmarking/Compiler.hs
@@ -18,6 +18,7 @@ import           Data.Text (Text)
 import qualified Data.Text as Text
 
 import           Cardano.Api
+import           Cardano.Benchmarking.Script.Action
 import           Cardano.Benchmarking.Script.Setters(Tag(..))
 import           Cardano.Benchmarking.Script.Store (KeyName, Name (..), WalletName)
 import           Cardano.Benchmarking.Script.Types
